Change structure for CMake compile
